Move pause/resume utilities to the engine
