+++ src/prefect/states.py
@@ -506,57 +506,3 @@ def Late(
         State: a Late state
     """
     return schemas.states.Late(cls=cls, scheduled_time=scheduled_time, **kwargs)
-
-
-@sync_compatible
-async def pause(timeout: int = 300, poll_interval: int = None):
-    import anyio
-    import time
-
-    from prefect.client.orion import get_client
-    from prefect.context import (
-        FlowRunContext,
-        TaskRunContext,
-    )
-    from prefect.logging.loggers import get_run_logger
-    from prefect.orion.schemas.states import Paused
-
-    poll_interval = poll_interval if poll_interval is not None else 10
-
-    if TaskRunContext.get():
-        raise RuntimeError("Can't pause task runs!")
-
-    frc = FlowRunContext.get()
-    logger = get_run_logger(context=frc)
-
-    logger.info("Pausing flow, execution will continue when this flow run is resumed.")
-    client = get_client()
-    response = await client.set_flow_run_state(
-        frc.flow_run.id,
-        Paused(),
-        force=True,
-    )
-
-    with anyio.move_on_after(timeout):
-        while True:
-            time.sleep(poll_interval)
-            flow_run = await client.read_flow_run(frc.flow_run.id)
-            if flow_run.state.is_running():
-                logger.info("Resuming flow run execution!")
-                return
-
-    raise FlowPauseTimeout("Flow run was paused and never resumed.")
-
-
-@sync_compatible
-async def resume(flow_run_id):
-    import pendulum
-
-    from prefect.client.orion import get_client
-    from prefect.orion.schemas.states import Running
-
-    client = get_client()
-    response = await client.set_flow_run_state(
-        flow_run_id,
-        Running(name="Resuming"),
-    )
